// frontend/src/openapi/models/WorkshopFunding.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FundingSourceEnum } from './FundingSourceEnum';
import {
    FundingSourceEnumFromJSON,
    FundingSourceEnumFromJSONTyped,
    FundingSourceEnumToJSON,
    FundingSourceEnumToJSONTyped,
} from './FundingSourceEnum';
import type { ChatGPTAccount } from './ChatGPTAccount';
import {
    ChatGPTAccountFromJSON,
    ChatGPTAccountFromJSONTyped,
    ChatGPTAccountToJSON,
    ChatGPTAccountToJSONTyped,
} from './ChatGPTAccount';

/**
 *
 * @export
 * @interface WorkshopFunding
 */
export interface WorkshopFunding {
    /**
     *
     * @type {boolean}
     * @memberof WorkshopFunding
     */
    enabled: boolean;
    /**
     *
     * @type {FundingSourceEnum}
     * @memberof WorkshopFunding
     */
    fundingSource: FundingSourceEnum;
    /**
     *
     * @type {string}
     * @memberof WorkshopFunding
     */
    accountId: string | null;
    /**
     *
     * @type {string}
     * @memberof WorkshopFunding
     */
    model: string;
    /**
     *
     * @type {string}
     * @memberof WorkshopFunding
     */
    usageUrl: string;
    /**
     *
     * @type {Array<ChatGPTAccount>}
     * @memberof WorkshopFunding
     */
    accounts: Array<ChatGPTAccount>;
}



/**
 * Check if a given object implements the WorkshopFunding interface.
 */
export function instanceOfWorkshopFunding(value: object): value is WorkshopFunding {
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    if (!('fundingSource' in value) || value['fundingSource'] === undefined) return false;
    if (!('accountId' in value) || value['accountId'] === undefined) return false;
    if (!('model' in value) || value['model'] === undefined) return false;
    if (!('usageUrl' in value) || value['usageUrl'] === undefined) return false;
    if (!('accounts' in value) || value['accounts'] === undefined) return false;
    return true;
}

export function WorkshopFundingFromJSON(json: any): WorkshopFunding {
    return WorkshopFundingFromJSONTyped(json, false);
}

export function WorkshopFundingFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkshopFunding {
    if (json == null) {
        return json;
    }
    return {

        'enabled': json['enabled'],
        'fundingSource': FundingSourceEnumFromJSON(json['funding_source']),
        'accountId': json['account_id'],
        'model': json['model'],
        'usageUrl': json['usage_url'],
        'accounts': ((json['accounts'] as Array<any>).map(ChatGPTAccountFromJSON)),
    };
}

export function WorkshopFundingToJSON(json: any): WorkshopFunding {
    return WorkshopFundingToJSONTyped(json, false);
}

export function WorkshopFundingToJSONTyped(value?: WorkshopFunding | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'enabled': value['enabled'],
        'funding_source': FundingSourceEnumToJSON(value['fundingSource']),
        'account_id': value['accountId'],
        'model': value['model'],
        'usage_url': value['usageUrl'],
        'accounts': ((value['accounts'] as Array<any>).map(ChatGPTAccountToJSON)),
    };
}